JBPM-4039: Create JAXB representation of the Audit/BAM interface (for REST/JMS)
- plus small fixes for the JaxbTaskData and JaxbComment classes

+++ kogito-jbpm/jbpm-audit/src/test/java/org/jbpm/process/audit/JPAWorkingMemoryDbLoggerTest.java
@@ -376,7 +376,7 @@ public class JPAWorkingMemoryDbLoggerTest {
         assertNotNull(processInstance.getEnd());
         assertEquals(processInstanceId, processInstance.getProcessInstanceId());
         assertEquals("com.sample.ruleflow", processInstance.getProcessId());
-        assertEquals(ProcessInstance.STATE_COMPLETED, processInstance.getStatus());
+        assertEquals(ProcessInstance.STATE_COMPLETED, processInstance.getStatus().intValue());
         List<NodeInstanceLog> nodeInstances = JPAProcessInstanceDbLog.findNodeInstances(processInstanceId);
         assertEquals(6, nodeInstances.size());
         for (NodeInstanceLog nodeInstance: nodeInstances) {
